[ICU-3905] Fix some compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/16760 Former-commit-id: 86cb03c5c66c11418fbbc8ec2d27c1a17c2865ee

+++ icu4c/source/test/intltest/svccoll.cpp
@@ -266,7 +266,7 @@ public:
   {
     CollatorInfo** p;
     for (p = info; *p; ++p) {}
-    count = p - info;
+    count = (int32_t)(p - info);
   }
 
   ~TestFactory() {
